+++ Objects/o_blob.py
