+++ commit_tree.py
@@ -1,4 +1,4 @@
-from Objects.commit import Commit
+from Objects.o_commit import Commit
 
 def commit_tree(tree_sha, message, parent_sha=None):
   commit = Commit()
