--- fichier.txt
@@ -1,1 +1,0 @@
-Ceci est un fichier de textekpepkroperkeprekeo
