--- test/objects/commit/aede5ac7a0b18f863ad1a3e6b44731dc279d0b96
@@ -1,3 +1,0 @@
-tree 3513413266ebbad20240bf88651e7745bea68d69
-message bonjour
-date 2025-06-18T14:11:06.590542
